FEAT: add API connection
- Main page API Connection
- Calendar page API Connection

+++ app/src/main/java/com/example/mulga/data/dto/response/DailyTransactionSummaryDto.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class DailyTransactionSummaryDto(
     val isValid: Boolean,
-    val income: Long,
-    val expense: Long
+    val income: Int,
+    val expense: Int
 )
