+++ app/src/main/java/com/example/mulga/domain/model/DailyTransactionSummaryEntity.kt
@@ -2,6 +2,6 @@ package com.example.mulga.domain.model
 
 data class DailyTransactionSummaryEntity(
     val isValid: Boolean,
-    val income: Long,
-    val expense: Long
+    val income: Int,
+    val expense: Int
 )
