+++ app/src/main/java/com/example/mulga/domain/model/MonthlyTransactionEntity.kt
@@ -1,7 +1,7 @@
 package com.example.mulga.domain.model
 
 data class MonthlyTransactionEntity(
-    val monthTotal: Long,
+    val monthTotal: Int,
     val year: Int,
     val month: Int,
     val dailySummaries: Map<Int, DailyTransactionSummaryEntity>,
